added parent suggestion for add page
