+++ apps/web/components/ui/PageModal/PageModal.vue
@@ -87,79 +87,6 @@
 import { SfIconClose, SfInput } from '@storefront-ui/vue';
 import Multiselect from 'vue-multiselect';
 import { ErrorMessage } from 'vee-validate';
-import { useForm, ErrorMessage } from 'vee-validate';
-import { toTypedSchema } from '@vee-validate/yup';
-import { object, string } from 'yup';
-import type { CategoryEntry, CategoryTreeItem } from '@plentymarkets/shop-api';
-import { categoryTreeGetters } from '@plentymarkets/shop-api';
-
-const router = useRouter();
-const { setCategoryId } = useCategoryIdHelper();
-const { pageModalOpen, togglePageModal, setSettingsCategory } = useSiteConfiguration();
-const { data: newCategory, addCategory } = useCategoryManagement();
-const { data, getCategories, addNewPageToTree } = useCategoriesSearch();
-
-const fetchCategoriesByName = async (name: string = '') => {
-  await getCategories({
-    type: 'in:item,content',
-    sortBy: 'position_asc,name_asc',
-    with: 'details,clients',
-    name: name ? `like:${name}` : '',
-  });
-};
-const loadInitialCategories = async () => {
-  await fetchCategoriesByName();
-};
-watch(
-  () => pageModalOpen.value,
-  async (isOpen) => {
-    if (isOpen) {
-      resetForm();
-      parentPage.value = emptyCategoryItem;
-      pageType.value = pageTypes.value[0];
-      await loadInitialCategories();
-    }
-  },
-);
-
-const validationSchema = toTypedSchema(
-  object({
-    pageName: string().required('Enter a page name').default(''),
-  }),
-);
-
-const { errors, meta, defineField, handleSubmit, resetForm } = useForm({
-  validationSchema: validationSchema,
-});
-
-const createNewPage = async () => {
-  if (!meta.value.valid) {
-    return;
-  }
-
-  await addCategory({
-    name: pageName?.value || '',
-    type: pageType.value.value,
-    parentCategoryId: categoryTreeGetters.getId(parentPage.value) || null,
-  });
-const { setCategoryId } = useCategoryIdHelper();
-
-  addNewPageToTree(newCategory.value);
-  await redirectToNewPage(newCategory.value);
-};
-
-const redirectToNewPage = async (newCategory: CategoryEntry) => {
-  await router.push({
-    path: newCategory.details[0].nameUrl,
-  });
-  setCategoryId(
-    newCategory.id,
-    newCategory.parentCategoryId,
-    newCategory.details[0].name,
-    newCategory.details[0].nameUrl,
-  );
-  setSettingsCategory({} as CategoryTreeItem, 'general-menu');
-};
 
 const {
   _isReady,
